@@ -3,11 +3,11 @@
     "description": "The ActivityPub protocol is a decentralized social networking protocol based upon the ActivityStreams 2.0 data format.",
     "type": "wordpress-plugin",
     "require": {
-        "php": ">=7.0",
+        "php": ">=7.2",
         "composer/installers": "^1.0 || ^2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.7.21 || ^6.5 || ^7.5 || ^8",
+        "phpunit/phpunit": "^8 || ^9",
         "phpcompatibility/php-compatibility": "*",
         "phpcompatibility/phpcompatibility-wp": "*",
         "squizlabs/php_codesniffer": "3.*",
@@ -39,6 +39,9 @@
             "composer install",
             "bin/install-wp-tests.sh activitypub-test root activitypub-test test-db latest true",
             "vendor/bin/phpunit"
+        ],
+        "test:wp-env": [
+            "wp-env run tests-cli --env-cwd=\"wp-content/plugins/activitypub\" vendor/bin/phpunit"
         ],
         "lint": [
             "vendor/bin/phpcs"
