@@ -10,7 +10,7 @@ namespace Activitypub\Rest;
 use WP_REST_Server;
 use WP_REST_Response;
 use Activitypub\Activity\Activity;
-use Activitypub\Collection\Actors as User_Collection;
+use Activitypub\Collection\Actors;
 
 use function Activitypub\get_context;
 use function Activitypub\url_to_authorid;
@@ -45,7 +45,7 @@ class Inbox {
 					'methods'             => WP_REST_Server::CREATABLE,
 					'callback'            => array( self::class, 'shared_inbox_post' ),
 					'args'                => self::shared_inbox_post_parameters(),
-					'permission_callback' => '__return_true',
+					'permission_callback' => array( 'Activitypub\Rest\Server', 'verify_signature' ),
 				),
 			)
 		);
@@ -58,13 +58,13 @@ class Inbox {
 					'methods'             => WP_REST_Server::CREATABLE,
 					'callback'            => array( self::class, 'user_inbox_post' ),
 					'args'                => self::user_inbox_post_parameters(),
-					'permission_callback' => '__return_true',
+					'permission_callback' => array( 'Activitypub\Rest\Server', 'verify_signature' ),
 				),
 				array(
 					'methods'             => WP_REST_Server::READABLE,
 					'callback'            => array( self::class, 'user_inbox_get' ),
 					'args'                => self::user_inbox_get_parameters(),
-					'permission_callback' => '__return_true',
+					'permission_callback' => array( 'Activitypub\Rest\Server', 'verify_signature' ),
 				),
 			)
 		);
@@ -78,14 +78,14 @@ class Inbox {
 	 */
 	public static function user_inbox_get( $request ) {
 		$user_id = $request->get_param( 'user_id' );
-		$user    = User_Collection::get_by_various( $user_id );
+		$user    = Actors::get_by_various( $user_id );
 
 		if ( is_wp_error( $user ) ) {
 			return $user;
 		}
 
 		/**
-		 * Action triggered prior to the ActivityPub profile being created and sent to the client.
+		 * Fires before the ActivityPub inbox is created and sent to the client.
 		 */
 		\do_action( 'activitypub_rest_inbox_pre' );
 
@@ -103,14 +103,14 @@ class Inbox {
 		// phpcs:enable WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
 
 		/**
-		 * Filter the ActivityPub inbox array.
+		 * Filters the ActivityPub inbox data before it is sent to the client.
 		 *
 		 * @param array $json The ActivityPub inbox array.
 		 */
 		$json = \apply_filters( 'activitypub_rest_inbox_array', $json );
 
 		/**
-		 * Action triggered after the ActivityPub profile has been created and sent to the client.
+		 * Fires after the ActivityPub inbox has been created and sent to the client.
 		 */
 		\do_action( 'activitypub_inbox_post' );
 
@@ -129,7 +129,7 @@ class Inbox {
 	 */
 	public static function user_inbox_post( $request ) {
 		$user_id = $request->get_param( 'user_id' );
-		$user    = User_Collection::get_by_various( $user_id );
+		$user    = Actors::get_by_various( $user_id );
 
 		if ( is_wp_error( $user ) ) {
 			return $user;
