@@ -79,7 +79,7 @@ abstract class Base {
 	}
 
 	/**
-	 * Transform the WordPress Object into an ActivityPub Object.
+	 * Transform the item into an ActivityPub Object.
 	 *
 	 * @return Base_Object|object The ActivityPub Object.
 	 */
@@ -87,6 +87,15 @@ abstract class Base {
 		$activitypub_object = new Base_Object();
 
 		return $this->transform_object_properties( $activitypub_object );
+	}
+
+	/**
+	 * Transform the item to an ActivityPub ID.
+	 *
+	 * @return string The ID of the WordPress Object.
+	 */
+	public function to_id() {
+		return $this->get_id();
 	}
 
 	/**
@@ -120,9 +129,20 @@ abstract class Base {
 
 	/**
 	 * Get the replies Collection.
+	 *
+	 * @return array The replies collection.
 	 */
 	public function get_replies() {
 		return Replies::get_collection( $this->wp_object );
+	}
+
+	/**
+	 * Returns the default media type for an Object.
+	 *
+	 * @return string The media type.
+	 */
+	public function get_media_type() {
+		return 'text/html';
 	}
 
 	/**
