@@ -13,6 +13,7 @@ use Activitypub\Model\Blog;
 use Activitypub\Collection\Actors;
 
 use function Activitypub\esc_hashtag;
+use function Activitypub\object_to_uri;
 use function Activitypub\is_single_user;
 use function Activitypub\get_enclosures;
 use function Activitypub\get_upload_baseurl;
@@ -151,7 +152,7 @@ class Post extends Base {
 
 		switch ( \get_post_status( $post ) ) {
 			case 'trash':
-				$permalink = \get_post_meta( $post->ID, 'activitypub_canonical_url', true );
+				$permalink = \get_post_meta( $post->ID, '_activitypub_canonical_url', true );
 				break;
 			case 'draft':
 				// Get_sample_permalink is in wp-admin, not always loaded.
@@ -208,7 +209,63 @@ class Post extends Base {
 		 */
 		$thumbnail = apply_filters(
 			'activitypub_get_image',
-			self::get_wordpress_attachment( $id, $image_size ),
+			$this->get_wordpress_attachment( $id, $image_size ),
+			$id,
+			$image_size
+		);
+
+		if ( ! $thumbnail ) {
+			return null;
+		}
+
+		$mime_type = \get_post_mime_type( $id );
+
+		$image = array(
+			'type'      => 'Image',
+			'url'       => \esc_url( $thumbnail[0] ),
+			'mediaType' => \esc_attr( $mime_type ),
+		);
+
+		$alt = \get_post_meta( $id, '_wp_attachment_image_alt', true );
+		if ( $alt ) {
+			$image['name'] = \wp_strip_all_tags( \html_entity_decode( $alt ) );
+		}
+
+		return $image;
+	}
+
+	/**
+	 * Returns an Icon, based on the Featured Image with a fallback to the site-icon.
+	 *
+	 * @return array|null The Icon or null if no icon is available.
+	 */
+	protected function get_icon() {
+		$post_id = $this->wp_object->ID;
+
+		// List post thumbnail first if this post has one.
+		if ( \has_post_thumbnail( $post_id ) ) {
+			$id = \get_post_thumbnail_id( $post_id );
+		} else {
+			// Try site_logo, falling back to site_icon, first.
+			$id = get_option( 'site_icon' );
+		}
+
+		if ( ! $id ) {
+			return null;
+		}
+
+		$image_size = 'thumbnail';
+
+		/**
+		 * Filter the image URL returned for each post.
+		 *
+		 * @param array|false $thumbnail  The image URL, or false if no image is available.
+		 * @param int         $id         The attachment ID.
+		 * @param string      $image_size The image size to retrieve. Set to 'large' by default.
+		 */
+		$thumbnail = apply_filters(
+			'activitypub_get_image',
+			$this->get_wordpress_attachment( $id, $image_size ),
 			$id,
 			$image_size
 		);
@@ -244,8 +301,15 @@ class Post extends Base {
 			return array();
 		}
 
-		// Once upon a time we only supported images, but we now support audio/video as well.
-		// We maintain the image-centric naming for backwards compatibility.
+		/**
+		 * Filters the maximum number of media attachments allowed in a post.
+		 *
+		 * Despite the name suggesting only images, this filter controls the maximum number
+		 * of all media attachments (images, audio, and video) that can be included in an
+		 * ActivityPub post. The name is maintained for backwards compatibility.
+		 *
+		 * @param int $max_media Maximum number of media attachments. Default ACTIVITYPUB_MAX_IMAGE_ATTACHMENTS.
+		 */
 		$max_media = \intval(
 			\apply_filters(
 				'activitypub_max_image_attachments',
@@ -254,9 +318,9 @@ class Post extends Base {
 		);
 
 		$media = array(
+			'image' => array(),
 			'audio' => array(),
 			'video' => array(),
-			'image' => array(),
 		);
 		$id    = $this->wp_object->ID;
 
@@ -273,7 +337,7 @@ class Post extends Base {
 			$media = $this->get_classic_editor_images( $media, $max_media );
 		}
 
-		$media      = self::filter_media_by_object_type( $media, \get_post_format( $this->wp_object ), $this->wp_object );
+		$media      = $this->filter_media_by_object_type( $media, \get_post_format( $this->wp_object ), $this->wp_object );
 		$unique_ids = \array_unique( \array_column( $media, 'id' ) );
 		$media      = \array_intersect_key( $media, $unique_ids );
 		$media      = \array_slice( $media, 0, $max_media );
@@ -288,7 +352,7 @@ class Post extends Base {
 		 */
 		$media = \apply_filters( 'activitypub_attachment_ids', $media, $this->wp_object );
 
-		$attachments = \array_filter( \array_map( array( self::class, 'wp_attachment_to_activity_attachment' ), $media ) );
+		$attachments = \array_filter( \array_map( array( $this, 'wp_attachment_to_activity_attachment' ), $media ) );
 
 		/**
 		 * Filter the attachments for a post.
@@ -318,14 +382,16 @@ class Post extends Base {
 		foreach ( $enclosures as $enclosure ) {
 			// Check if URL is an attachment.
 			$attachment_id = \attachment_url_to_postid( $enclosure['url'] );
+
 			if ( $attachment_id ) {
 				$enclosure['id']        = $attachment_id;
 				$enclosure['url']       = \wp_get_attachment_url( $attachment_id );
 				$enclosure['mediaType'] = \get_post_mime_type( $attachment_id );
 			}
 
-			$mime_type       = $enclosure['mediaType'];
-			$mime_type_parts = \explode( '/', $mime_type );
+			$mime_type         = $enclosure['mediaType'];
+			$mime_type_parts   = \explode( '/', $mime_type );
+			$enclosure['type'] = \ucfirst( $mime_type_parts[0] );
 
 			switch ( $mime_type_parts[0] ) {
 				case 'image':
@@ -359,7 +425,7 @@ class Post extends Base {
 
 		$blocks = \parse_blocks( $this->wp_object->post_content );
 
-		return self::get_media_from_blocks( $blocks, $media );
+		return $this->get_media_from_blocks( $blocks, $media );
 	}
 
 	/**
@@ -370,11 +436,11 @@ class Post extends Base {
 	 *
 	 * @return array The image IDs.
 	 */
-	protected static function get_media_from_blocks( $blocks, $media ) {
+	protected function get_media_from_blocks( $blocks, $media ) {
 		foreach ( $blocks as $block ) {
 			// Recurse into inner blocks.
 			if ( ! empty( $block['innerBlocks'] ) ) {
-				$media = self::get_media_from_blocks( $block['innerBlocks'], $media );
+				$media = $this->get_media_from_blocks( $block['innerBlocks'], $media );
 			}
 
 			switch ( $block['blockName'] ) {
@@ -388,10 +454,21 @@ class Post extends Base {
 							$alt = $match[2];
 						}
 
-						$media['image'][] = array(
-							'id'  => $block['attrs']['id'],
-							'alt' => $alt,
-						);
+						$found = false;
+						foreach ( $media['image'] as $i => $image ) {
+							if ( $image['id'] === $block['attrs']['id'] ) {
+								$media['image'][ $i ]['alt'] = $alt;
+								$found                       = true;
+								break;
+							}
+						}
+
+						if ( ! $found ) {
+							$media['image'][] = array(
+								'id'  => $block['attrs']['id'],
+								'alt' => $alt,
+							);
+						}
 					}
 					break;
 				case 'core/audio':
@@ -585,7 +662,7 @@ class Post extends Base {
 	 *
 	 * @return array The filtered media IDs.
 	 */
-	protected static function filter_media_by_object_type( $media, $type, $wp_object ) {
+	protected function filter_media_by_object_type( $media, $type, $wp_object ) {
 		/**
 		 * Filter the object type for media attachments.
 		 *
@@ -610,7 +687,7 @@ class Post extends Base {
 	 *
 	 * @return array The ActivityPub Attachment.
 	 */
-	public static function wp_attachment_to_activity_attachment( $media ) {
+	public function wp_attachment_to_activity_attachment( $media ) {
 		if ( ! isset( $media['id'] ) ) {
 			return $media;
 		}
@@ -633,7 +710,7 @@ class Post extends Base {
 				 */
 				$thumbnail = apply_filters(
 					'activitypub_get_image',
-					self::get_wordpress_attachment( $id, $image_size ),
+					$this->get_wordpress_attachment( $id, $image_size ),
 					$id,
 					$image_size
 				);
@@ -672,7 +749,11 @@ class Post extends Base {
 					$attachment['width']  = \esc_attr( $meta['width'] );
 					$attachment['height'] = \esc_attr( $meta['height'] );
 				}
-				// @todo: add `icon` support for audio/video attachments. Maybe use post thumbnail?
+
+				if ( $this->get_icon() ) {
+					$attachment['icon'] = object_to_uri( $this->get_icon() );
+				}
+
 				break;
 		}
 
@@ -695,7 +776,7 @@ class Post extends Base {
 	 *
 	 * @return array|false Array of image data, or boolean false if no image is available.
 	 */
-	protected static function get_wordpress_attachment( $id, $image_size = 'large' ) {
+	protected function get_wordpress_attachment( $id, $image_size = 'large' ) {
 		/**
 		 * Hook into the image retrieval process. Before image retrieval.
 		 *
@@ -918,7 +999,7 @@ class Post extends Base {
 		 */
 		do_action( 'activitypub_before_get_content', $post );
 
-		add_filter( 'render_block_core/embed', array( self::class, 'revert_embed_links' ), 10, 2 );
+		add_filter( 'render_block_core/embed', array( $this, 'revert_embed_links' ), 10, 2 );
 
 		// phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
 		$post    = $this->wp_object;
@@ -940,6 +1021,12 @@ class Post extends Base {
 		$content = \preg_replace( '/[\n\r\t]/', '', $content );
 		$content = \trim( $content );
 
+		/**
+		 * Filters the post content before it is transformed for ActivityPub.
+		 *
+		 * @param string  $content The post content to be transformed.
+		 * @param WP_Post $post    The post object being transformed.
+		 */
 		$content = \apply_filters( 'activitypub_the_content', $content, $post );
 
 		// Don't need these anymore, should never appear in a post.
@@ -969,6 +1056,17 @@ class Post extends Base {
 			$template .= '[ap_content]';
 		}
 
+		/**
+		 * Filters the template used to generate ActivityPub object content.
+		 *
+		 * This filter allows developers to modify the template that determines how post
+		 * content is formatted in ActivityPub objects. The template can include special
+		 * shortcodes like [ap_title] and [ap_content] that are processed during content
+		 * generation.
+		 *
+		 * @param string  $template  The template string containing shortcodes.
+		 * @param WP_Post $wp_object The WordPress post object being transformed.
+		 */
 		return apply_filters( 'activitypub_object_content_template', $template, $this->wp_object );
 	}
 
@@ -1027,7 +1125,7 @@ class Post extends Base {
 		$blocks = \parse_blocks( $this->wp_object->post_content );
 
 		foreach ( $blocks as $block ) {
-			if ( 'activitypub/reply' === $block['blockName'] ) {
+			if ( 'activitypub/reply' === $block['blockName'] && isset( $block['attrs']['url'] ) ) {
 				// We only support one reply block per post for now.
 				return $block['attrs']['url'];
 			}
@@ -1117,7 +1215,7 @@ class Post extends Base {
 	 *
 	 * @return string A block level link
 	 */
-	public static function revert_embed_links( $block_content, $block ) {
+	public function revert_embed_links( $block_content, $block ) {
 		if ( ! isset( $block['attrs']['url'] ) ) {
 			return $block_content;
 		}
