@@ -1,12 +1,11 @@
-
 import apiFetch from '@wordpress/api-fetch';
 import { useEffect, useState } from '@wordpress/element';
 import { Button, Modal } from '@wordpress/components';
 import { __, sprintf } from '@wordpress/i18n';
 import { ButtonStyle, getPopupStyles } from './button-style';
 import { Dialog } from '../shared/dialog';
+import { useOptions } from '../shared/use-options';
 import './style.scss';
-const { namespace } = window._activityPubOptions;
 
 const DEFAULT_PROFILE_DATA = {
 	avatar: '',
@@ -25,6 +24,7 @@ function getNormalizedProfile( profile ) {
 }
 
 function fetchProfile( userId ) {
+	const { namespace } = useOptions();
 	const fetchOptions = {
 		headers: { Accept: 'application/activity+json' },
 		path: `/${ namespace }/actors/${ userId }`,
@@ -54,7 +54,13 @@ function Follow( { profile, popupStyles, userId } ) {
 
 	return (
 		<>
-			<Button className="activitypub-profile__follow" onClick={ () => setIsOpen( true ) } >
+			<Button
+				className="activitypub-profile__follow"
+				onClick={ () => setIsOpen( true ) }
+				aria-haspopup="dialog"
+				aria-expanded={ isOpen }
+				aria-label={  __( 'Follow me on the Fediverse', 'activitypub' ) }
+			>
 				{ __( 'Follow', 'activitypub' ) }
 			</Button>
 			{ isOpen && (
@@ -62,6 +68,8 @@ function Follow( { profile, popupStyles, userId } ) {
 				className="activitypub-profile__confirm activitypub__modal"
 				onRequestClose={ () => setIsOpen( false ) }
 				title={ title }
+				aria-label={ title }
+				role="dialog"
 				>
 					<DialogFollow profile={ profile } userId={ userId } />
 					<style>{ popupStyles }</style>
@@ -72,6 +80,7 @@ function Follow( { profile, popupStyles, userId } ) {
 }
 
 function DialogFollow( { profile, userId } ) {
+	const { namespace } = useOptions();
 	const { webfinger } = profile;
 	const actionText = __( 'Follow', 'activitypub' );
 	const resourceUrl = `/${ namespace }/actors/${userId}/remote-follow?resource=`;
