@@ -4,11 +4,11 @@ import { __, sprintf } from '@wordpress/i18n';
 import { Icon, cancelCircleFilled } from '@wordpress/icons';
 import { Dialog } from '../shared/dialog';
 import { useRemoteUser } from '../shared/use-remote-user';
+import { useOptions } from '../shared/use-options';
 import './style.scss';
-const { namespace } = window._activityPubOptions;
-
 
 function DialogReply( { selectedComment, commentId } ) {
+	const { namespace } = useOptions();
 	const actionText = __( 'Reply', 'activitypub' );
 	const resourceUrl = `/${ namespace }/comments/${commentId}/remote-reply?resource=`;
 	const copyDescription = __( 'Copy and paste the Comment URL into the search field of your favorite fediverse app or server.', 'activitypub' );
