@@ -62,36 +62,44 @@ export function Dialog( { actionText, copyDescription, handle, resourceUrl, myPr
 	}, [ remoteProfile ] );
 
 	return (
-		<div className="activitypub__dialog">
+		<div className="activitypub__dialog" role="dialog" aria-labelledby="dialog-title">
 			<div className="activitypub-dialog__section">
-				<h4>{ myProfileHeader }</h4>
-				<div className="activitypub-dialog__description">
+				<h4 id="dialog-title">{ myProfileHeader }</h4>
+				<div className="activitypub-dialog__description" id="copy-description">
 					{ copyDescription }
 				</div>
 				<div className="activitypub-dialog__button-group">
-					<input type="text" value={ handle } readOnly />
-					<Button ref={ ref }>
+					<label htmlFor="profile-handle" className="screen-reader-text">
+						{ copyDescription }
+					</label>
+					<input type="text" id="profile-handle" value={ handle } readOnly />
+					<Button ref={ ref } aria-label={ __( 'Copy handle to clipboard', 'activitypub' ) }>
 						<Icon icon={ buttonIcon } />
 						{ __( 'Copy', 'activitypub' ) }
 					</Button>
 				</div>
 			</div>
 			<div className="activitypub-dialog__section">
-				<h4>{ __( 'Your Profile', 'activitypub' ) }</h4>
-				<div className="activitypub-dialog__description">
+				<h4 id="remote-profile-title">{ __( 'Your Profile', 'activitypub' ) }</h4>
+				<div className="activitypub-dialog__description" id="remote-profile-description">
 					{ createInterpolateElement(
 						__( 'Or, if you know your own profile, we can start things that way! (eg <code>@yourusername@example.com</code>)', 'activitypub' ),
 						{ code: <code /> }
 					) }
 				</div>
 				<div className="activitypub-dialog__button-group">
+					<label htmlFor="remote-profile" className="screen-reader-text">
+						{ __( 'Enter your ActivityPub profile', 'activitypub' ) }
+					</label>
 					<input
 						type="text"
+						id="remote-profile"
 						value={ remoteProfile }
 						onKeyDown={ ( event ) => { event?.code === 'Enter' && retrieveAndFollow() } }
 						onChange={ e => setRemoteProfile( e.target.value ) }
+						aria-invalid={ buttonText === invalidText }
 					/>
-					<Button onClick={ retrieveAndFollow }>
+					<Button onClick={ retrieveAndFollow } aria-label={ __( 'Submit profile', 'activitypub' ) }>
 						<Icon icon={ people } />
 						{ buttonText }
 					</Button>
