@@ -34,7 +34,7 @@
 								</label>
 							</p>
 							<p class="description">
-								<?php echo \wp_kses( \__( 'Every author on this blog (with the <code>activitypub</code> capability) gets their own ActivityPub profile.', 'activitypub' ), array( 'code' => array() ) ); ?>
+								<?php echo \wp_kses( \__( 'Every user with an account on this blog and the <code>activitypub</code> capability enabled gets their own ActivityPub profile.', 'activitypub' ), array( 'code' => array() ) ); ?>
 								<?php // translators: %s is a URL. ?>
 								<strong><?php echo \wp_kses( sprintf( \__( 'You can add/remove the capability in the <a href="%s">user settings.</a>', 'activitypub' ), admin_url( '/users.php' ) ), array( 'a' => array( 'href' => array() ) ) ); ?></strong>
 								<?php echo \wp_kses( \__( 'Select all the users you want to update, choose the method from the drop-down list and click on the "Apply" button.', 'activitypub' ), array( 'code' => array() ) ); ?>
@@ -59,10 +59,9 @@
 							</p>
 						</td>
 					</tr>
+					<?php \do_settings_fields( 'activitypub', 'user' ); ?>
 				</tbody>
 			</table>
-
-			<?php \do_settings_fields( 'activitypub', 'user' ); ?>
 		</div>
 
 		<div class="box">
@@ -112,7 +111,7 @@
 											<li><code>[ap_content]</code> - <?php \esc_html_e( 'The post&#8217;s content.', 'activitypub' ); ?></li>
 											<li><code>[ap_excerpt]</code> - <?php \esc_html_e( 'The post&#8217;s excerpt (may be truncated).', 'activitypub' ); ?></li>
 											<li><code>[ap_permalink]</code> - <?php \esc_html_e( 'The post&#8217;s permalink.', 'activitypub' ); ?></li>
-											<li><code>[ap_shortlink]</code> - <?php echo \wp_kses( \__( 'The pos&#8217;s shortlink. I can recommend <a href="https://wordpress.org/plugins/hum/" target="_blank">Hum</a>.', 'activitypub' ), 'default' ); ?></li>
+											<li><code>[ap_shortlink]</code> - <?php echo \wp_kses( \__( 'The post&#8217;s shortlink. I can recommend <a href="https://wordpress.org/plugins/hum/" target="_blank">Hum</a>.', 'activitypub' ), 'default' ); ?></li>
 											<li><code>[ap_hashtags]</code> - <?php \esc_html_e( 'The post&#8217;s tags as hashtags.', 'activitypub' ); ?></li>
 										</ul>
 										<p><?php \esc_html_e( 'You can find the full list with all possible attributes in the help section on the top-right of the screen.', 'activitypub' ); ?></p>
@@ -183,12 +182,47 @@
 							</p>
 						</td>
 					</tr>
+					<?php \do_settings_fields( 'activitypub', 'activity' ); ?>
 				</tbody>
 			</table>
-
-			<?php \do_settings_fields( 'activitypub', 'activity' ); ?>
 		</div>
-
+		<div class="box">
+			<h3><?php \esc_html_e( 'Notifications', 'activitypub' ); ?></h3>
+			<p><?php \esc_html_e( 'Choose which notifications you want to receive. The plugin currently only supports e-mail notifications, but we will add more options in the future.', 'activitypub' ); ?></p>
+			<table class="form-table">
+				<tbody>
+					<tr>
+						<th scope="row">
+							<?php \esc_html_e( 'Type', 'activitypub' ); ?>
+						</th>
+						<th>
+							<?php \esc_html_e( 'E-Mail', 'activitypub' ); ?>
+						</th>
+					</tr>
+					<tr>
+						<td scope="row">
+							<?php \esc_html_e( 'New followers', 'activitypub' ); ?>
+						</td>
+						<td>
+							<label>
+								<input type="checkbox" name="activitypub_mailer_new_follower" id="activitypub_mailer_new_follower" value="1" <?php \checked( '1', \get_option( 'activitypub_mailer_new_follower', '0' ) ); ?> />
+							</label>
+						</td>
+					</tr>
+					<tr>
+						<td scope="row">
+							<?php \esc_html_e( 'Direct Messages', 'activitypub' ); ?>
+						</td>
+						<td>
+							<label>
+								<input type="checkbox" name="activitypub_mailer_new_dm" id="activitypub_mailer_new_dm" value="1" <?php \checked( '1', \get_option( 'activitypub_mailer_new_dm', '0' ) ); ?> />
+							</label>
+						</td>
+					</tr>
+					<?php \do_settings_fields( 'activitypub', 'security' ); ?>
+				</tbody>
+			</table>
+		</div>
 		<div class="box">
 			<h3><?php \esc_html_e( 'General', 'activitypub' ); ?></h3>
 			<table class="form-table">
@@ -213,6 +247,36 @@
 							</p>
 						</td>
 					</tr>
+					<?php \do_settings_fields( 'activitypub', 'general' ); ?>
+					<?php \do_settings_fields( 'activitypub', 'server' ); ?>
+				</tbody>
+			</table>
+		</div>
+		<div class="box">
+			<h3><?php \esc_html_e( 'Security', 'activitypub' ); ?></h3>
+			<table class="form-table">
+				<tbody>
+					<?php if ( ! defined( 'ACTIVITYPUB_AUTHORIZED_FETCH' ) ) : ?>
+					<tr>
+						<th scope="row">
+							<?php \esc_html_e( 'Authorized-Fetch', 'activitypub' ); ?>
+						</th>
+						<td>
+							<p>
+								<label>
+									<input type="checkbox" name="activitypub_authorized_fetch" id="activitypub_authorized_fetch" value="1" <?php \checked( '1', \get_option( 'activitypub_authorized_fetch', '0' ) ); ?> />
+									<?php \esc_html_e( 'Require HTTP signature authentication on ActivityPub representations of public posts and profiles.', 'activitypub' ); ?>
+								</label>
+							</p>
+							<p class="description">
+								<?php \esc_html_e( '⚠ Secure mode has its limitations, which is why it is not enabled by default. It is not fully supported by all software in the fediverse, and some features may break, especially when interacting with Mastodon servers older than version 3.0. Additionally, since it requires authentication for public content, caching is not possible, leading to higher computational costs.', 'activitypub' ); ?>
+							</p>
+							<p class="description">
+								<?php \esc_html_e( '⚠ Secure mode does not hide the HTML representations of public posts and profiles. While HTML is a less consistant format (that potentially changes often) compared to first-class ActivityPub representations or the REST API, it still poses a potential risk for content scraping.', 'activitypub' ); ?>
+							</p>
+						</td>
+					</tr>
+					<?php endif; ?>
 					<tr>
 						<th scope="row">
 							<?php \esc_html_e( 'Blocklist', 'activitypub' ); ?>
@@ -232,10 +296,9 @@
 							</p>
 						</td>
 					</tr>
+					<?php \do_settings_fields( 'activitypub', 'security' ); ?>
 				</tbody>
 			</table>
-			<?php \do_settings_fields( 'activitypub', 'general' ); ?>
-			<?php \do_settings_fields( 'activitypub', 'server' ); ?>
 		</div>
 		<?php \do_settings_sections( 'activitypub' ); ?>
 
