@@ -5,9 +5,9 @@
  * @package Activitypub
  */
 
-\define( 'ACTIVITYPUB_DISABLE_REACTIONS', false );
 \define( 'ACTIVITYPUB_DISABLE_INCOMING_INTERACTIONS', false );
 
+\define( 'AP_TESTS_DIR', __DIR__ );
 $_tests_dir = \getenv( 'WP_TESTS_DIR' );
 
 if ( ! $_tests_dir ) {
@@ -34,8 +34,71 @@ function _manually_load_plugin() {
 }
 \tests_add_filter( 'muplugins_loaded', '_manually_load_plugin' );
 
+/**
+ * Disable HTTP requests.
+ *
+ * @param mixed  $response The value to return instead of making a HTTP request.
+ * @param array  $args     Request arguments.
+ * @param string $url      The request URL.
+ * @return mixed|false|WP_Error
+ */
+function http_disable_request( $response, $args, $url ) {
+	if ( false !== $response ) {
+		// Another filter has already overridden this request.
+		return $response;
+	}
+
+	/**
+	 * Allow HTTP requests to be made.
+	 *
+	 * @param bool  $allow Whether to allow the HTTP request.
+	 * @param array $args  Request arguments.
+	 * @param string $url  The request URL.
+	 */
+	if ( apply_filters( 'tests_allow_http_request', false, $args, $url ) ) {
+		// This request has been specifically permitted.
+		return false;
+	}
+
+	$backtrace = array_reverse( debug_backtrace() ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_debug_backtrace,PHPCompatibility.FunctionUse.ArgumentFunctionsReportCurrentValue.NeedsInspection
+	$trace_str = '';
+	foreach ( $backtrace as $frame ) {
+		if (
+			( isset( $frame['file'] ) && strpos( $frame['file'], 'phpunit.php' ) !== false ) ||
+			( isset( $frame['file'] ) && strpos( $frame['file'], 'wp-includes/http.php' ) !== false ) ||
+			( isset( $frame['file'] ) && strpos( $frame['file'], 'wp-includes/class-wp-hook.php' ) !== false ) ||
+			( isset( $frame['function'] ) && __FUNCTION__ === $frame['function'] ) ||
+			( isset( $frame['function'] ) && 'apply_filters' === $frame['function'] )
+		) {
+			continue;
+		}
+
+		if ( $trace_str ) {
+			$trace_str .= ', ';
+		}
+
+		if ( ! empty( $frame['file'] ) && ! empty( $frame['line'] ) ) {
+			$trace_str .= basename( $frame['file'] ) . ':' . $frame['line'];
+			if ( ! empty( $frame['function'] ) ) {
+				$trace_str .= ' ';
+			}
+		}
+
+		if ( ! empty( $frame['function'] ) ) {
+			if ( ! empty( $frame['class'] ) ) {
+				$trace_str .= $frame['class'] . '::';
+			}
+			$trace_str .= $frame['function'] . '()';
+		}
+	}
+
+	return new WP_Error( 'cancelled', 'Live HTTP request cancelled by bootstrap.php' );
+}
+\tests_add_filter( 'pre_http_request', 'http_disable_request', 99, 3 );
+
 // Start up the WP testing environment.
 require $_tests_dir . '/includes/bootstrap.php';
 require __DIR__ . '/class-activitypub-testcase-cache-http.php';
+require __DIR__ . '/class-test-rest-controller-testcase.php';
 
 \Activitypub\Migration::add_default_settings();
